Util for Path -> ContentBody. Test for repeatable.
